Recebendo valor double para fazer operações

+++ src/main/java/service/ClientService.java
@@ -42,7 +42,7 @@ public class ClientService {
         double value;
         do {
             System.out.println("Digite o valor para" + mensagem);
-            value = scanner.nextInt();
+            value = scanner.nextDouble();
             scanner.nextLine();
             if(value < 0){
                 System.out.println("valor Invalido tente novamente");
